Renamed and split up components
