+++ src/index.ts
@@ -107,7 +107,7 @@ export { default as Select } from './components/Select'
 export { default as Stepper } from './components/Stepper'
 export { default as Tab } from './components/Tab'
 export { default as Tag } from './components/Tag'
-export { default as TextCarousel } from './components/TextCarousel'
+export { default as TipCarousel } from './components/Carousel'
 export { default as Token } from './components/Token'
 export { default as UserCard } from './components/UserCard'
 // Hooks
